+++ src/providers/digitalocean.rs
@@ -10,16 +10,26 @@
  */
 
 use std::{
-    net::{Ipv4Addr, Ipv6Addr},
+    hash::{Hash, Hasher},
     time::Duration,
 };
 
-use crate::{http::HttpClientBuilder, strip_origin_from_name, DnsRecord, Error, IntoFqdn};
+use crate::{
+    http::HttpClientBuilder, strip_origin_from_name, ApiCacheFetcher, ApiCacheManager, DnsRecord,
+    DnsRecordTrait, Error, IntoFqdn,
+};
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone)]
 pub struct DigitalOceanProvider {
     client: HttpClientBuilder,
+    record_cache: ApiCacheManager<i64>,
+}
+
+struct DigitalOceanRecordFetcher<'a> {
+    client: &'a HttpClientBuilder,
+    name: &'a str,
+    domain: &'a str,
 }
 
 #[derive(Deserialize, Serialize, Clone, Debug)]
@@ -45,34 +55,16 @@ pub struct DomainRecord {
 }
 
 #[derive(Deserialize, Serialize, Clone, Debug)]
-#[serde(tag = "type")]
-#[allow(clippy::upper_case_acronyms)]
-pub enum RecordData {
-    A {
-        data: Ipv4Addr,
-    },
-    AAAA {
-        data: Ipv6Addr,
-    },
-    CNAME {
-        data: String,
-    },
-    NS {
-        data: String,
-    },
-    MX {
-        data: String,
-        priority: u16,
-    },
-    TXT {
-        data: String,
-    },
-    SRV {
-        data: String,
-        priority: u16,
-        port: u16,
-        weight: u16,
-    },
+pub struct RecordData {
+    #[serde(rename = "type")]
+    rr_type: String,
+    data: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    priority: Option<u16>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    port: Option<u16>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    weight: Option<u16>,
 }
 
 #[derive(Serialize, Debug)]
@@ -85,7 +77,10 @@ impl DigitalOceanProvider {
         let client = HttpClientBuilder::default()
             .with_header("Authorization", format!("Bearer {}", auth_token.as_ref()))
             .with_timeout(timeout);
-        Self { client }
+        Self {
+            client,
+            record_cache: ApiCacheManager::default(),
+        }
     }
 
     pub(crate) async fn create(
@@ -123,7 +118,14 @@ impl DigitalOceanProvider {
         let name = name.into_name();
         let domain = origin.into_name();
         let subdomain = strip_origin_from_name(&name, &domain);
-        let record_id = self.obtain_record_id(&name, &domain).await?;
+        let record_id = self
+            .record_cache
+            .get_or_update(&mut DigitalOceanRecordFetcher {
+                client: &self.client,
+                name: name.as_ref(),
+                domain: domain.as_ref(),
+            })
+            .await?;
 
         self.client
             .put(format!(
@@ -146,7 +148,14 @@ impl DigitalOceanProvider {
     ) -> crate::Result<()> {
         let name = name.into_name();
         let domain = origin.into_name();
-        let record_id = self.obtain_record_id(&name, &domain).await?;
+        let record_id = self
+            .record_cache
+            .get_or_update(&mut DigitalOceanRecordFetcher {
+                client: &self.client,
+                name: name.as_ref(),
+                domain: domain.as_ref(),
+            })
+            .await?;
 
         self.client
             .delete(format!(
@@ -156,13 +165,16 @@ impl DigitalOceanProvider {
             .await
             .map(|_| ())
     }
+}
 
-    async fn obtain_record_id(&self, name: &str, domain: &str) -> crate::Result<i64> {
-        let subdomain = strip_origin_from_name(name, domain);
+impl<'a> ApiCacheFetcher<i64> for DigitalOceanRecordFetcher<'a> {
+    async fn fetch_api_response(&mut self) -> crate::Result<i64> {
+        let subdomain = strip_origin_from_name(self.name, self.domain);
         self.client
             .get(format!(
-                "https://api.digitalocean.com/v2/domains/{domain}/records?{}",
-                Query::name(name).serialize()
+                "https://api.digitalocean.com/v2/domains/{}/records?{}",
+                self.domain,
+                Query::name(self.name).serialize()
             ))
             .send_with_retry::<ListDomainRecord>(3)
             .await
@@ -174,6 +186,13 @@ impl DigitalOceanProvider {
                     .map(|record| record.id)
                     .ok_or_else(|| Error::Api(format!("DNS Record {} not found", subdomain)))
             })
+    }
+}
+
+impl Hash for DigitalOceanRecordFetcher<'_> {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.name.hash(state);
+        self.domain.hash(state);
     }
 }
 
@@ -189,27 +208,12 @@ impl<'a> Query<'a> {
 
 impl From<DnsRecord> for RecordData {
     fn from(record: DnsRecord) -> Self {
-        match record {
-            DnsRecord::A { content } => RecordData::A { data: content },
-            DnsRecord::AAAA { content } => RecordData::AAAA { data: content },
-            DnsRecord::CNAME { content } => RecordData::CNAME { data: content },
-            DnsRecord::NS { content } => RecordData::NS { data: content },
-            DnsRecord::MX { content, priority } => RecordData::MX {
-                data: content,
-                priority,
-            },
-            DnsRecord::TXT { content } => RecordData::TXT { data: content },
-            DnsRecord::SRV {
-                content,
-                priority,
-                weight,
-                port,
-            } => RecordData::SRV {
-                data: content,
-                priority,
-                weight,
-                port,
-            },
+        RecordData {
+            rr_type: record.get_type().to_string(),
+            data: record.get_content(),
+            priority: record.get_priority(),
+            weight: record.get_weight(),
+            port: record.get_port(),
         }
     }
 }