+++ src/tests/desec_tests.rs
@@ -1,3 +1,14 @@
+/*
+ * Copyright Stalwart Labs LLC See the COPYING
+ * file at the top-level directory of this distribution.
+ *
+ * Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+ * https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+ * <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
+ * option. This file may not be copied, modified, or distributed
+ * except according to those terms.
+ */
+
 #[cfg(test)]
 mod tests {
     use crate::providers::desec::DesecDnsRecordRepresentation;
@@ -205,9 +216,15 @@ mod tests {
     #[tokio::test]
     #[ignore = "Requires desec API Token and domain configuration"]
     async fn integration_test() {
-        let token = ""; // <-- Fill in your deSEC API token here
-        let origin = ""; // <-- Fill in your domain (e.g., "example.com")
-        let domain = ""; // <-- Fill in your test subdomain (e.g., "test.example.com")
+        let token = std::env::var("DESEC_TOKEN").unwrap_or_else(
+            |_v| "".to_string(), /* <-- Fill in your deSEC API token here */
+        );
+        let origin = std::env::var("DESEC_ORIGIN").unwrap_or_else(
+            |_v| "".to_string(), /* <-- Fill in your domain (e.g., "example.com") */
+        );
+        let domain = std::env::var("DESEC_DOMAIN").unwrap_or_else(
+            |_v| "".to_string(), /* <-- Fill in your test subdomain (e.g., "test.example.com")*/
+        );
 
         assert!(
             !token.is_empty(),
@@ -222,17 +239,17 @@ mod tests {
             "Please configure your test subdomain in the integration test"
         );
 
-        let provider = DesecProvider::new(token, Some(Duration::from_secs(30)));
+        let provider = DesecProvider::new(&token, Some(Duration::from_secs(30)));
 
         // check creation
         let creation_result = provider
             .create(
-                domain,
+                &domain,
                 DnsRecord::A {
                     content: "1.1.1.1".parse().unwrap(),
                 },
                 3600,
-                origin,
+                &origin,
             )
             .await;
 
@@ -241,19 +258,19 @@ mod tests {
         // check modification
         let update_result = provider
             .update(
-                domain,
+                &domain,
                 DnsRecord::A {
                     content: "2.2.2.2".parse().unwrap(),
                 },
                 3600,
-                origin,
+                &origin,
             )
             .await;
 
         assert!(update_result.is_ok());
 
         // check deletion
-        let deletion_result = provider.delete(domain, origin, DnsRecordType::A).await;
+        let deletion_result = provider.delete(&domain, &origin, DnsRecordType::A).await;
 
         assert!(deletion_result.is_ok());
     }
@@ -278,7 +295,7 @@ mod tests {
             content: "test".to_string(),
         };
         let desec_record: DesecDnsRecordRepresentation = record.into();
-        assert_eq!(desec_record.content, "test");
+        assert_eq!(desec_record.content, "\"test\"");
         assert_eq!(desec_record.record_type, "TXT");
 
         let record = DnsRecord::MX {